backend/pyproject.toml
[project]
name = "feishu-ai-nav"
version = "1.0.0"
description = "飞书AI工具导航后端"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "greenlet>=3.0.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.26.0",
    "apscheduler>=3.10.4",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
postgres = ["asyncpg>=0.29.0"]

[tool.uv]
dev-dependencies = []
